+++ src/app/controls/handlers/feature-info-control.handler.spec.ts
@@ -2,6 +2,8 @@ import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { TestBed } from '@angular/core/testing';
 
 import { AppCfg, AppTasks } from '@api/model/app-cfg';
+import { TranslateModule } from '@ngx-translate/core';
+
 
 import { FeatureInfoControlHandler } from './feature-info-control.handler';
 import { AppConfigService } from '../../services/app-config.service';
@@ -30,7 +32,7 @@ describe('FeatureInfoControlHandler', () => {
     > as jest.Mocked<AppConfigService>;
 
     TestBed.configureTestingModule({
-      imports: [HttpClientTestingModule],
+      imports: [HttpClientTestingModule, TranslateModule.forRoot()],
       providers: [
         FeatureInfoControlHandler,
         { provide: SitnaApiService, useValue: mockSitnaApi },